refactor(runtime): 拆分 runtime 层为 state/managers/infra 三个子目录
- state/: ObservableState, VideoState, AppState
- managers/: ConfigManager, AccountManager, ThemeManager
- infra/: log_utils, resources
- 更新所有跨模块导入路径
