+++ src/danmaku_sender/runtime/managers/__init__.py
@@ -1,0 +1,5 @@
+from .config_manager import ConfigManager
+from .account_manager import AccountManager
+from .theme_manager import ThemeManager, Palette
+
+__all__ = ["ConfigManager", "AccountManager", "ThemeManager", "Palette"]
