+++ src/danmaku_sender/ui/monitor_page.py
@@ -11,7 +11,7 @@ from PySide6.QtWidgets import (
 )
 
 from danmaku_sender.types.models.common import MonitorStats, VideoTarget
-from danmaku_sender.runtime.app_state import AppState
+from danmaku_sender.runtime.state.app_state import AppState
 from danmaku_sender.repo.history_manager import HistoryManager
 
 from danmaku_sender.controller.monitor_controller import MonitorController
